pyproject.toml
[project]
name = "instagram-clone"
version = "0.1.0"
description = ""
authors = [
    {name = "Vlad_the_programmer", email = "vladyslavklymchuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref==3.8.1",
    "async-timeout==4.0.3",
    "attrs==24.2.0",
    "autobahn==24.4.2",
    "Automat==24.8.1",
    "certifi==2024.2.2",
    "cffi==1.16.0",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "charset-normalizer==3.3.2",
    "colorclass==2.2.2",
    "constantly==23.10.4",
    "coreapi==2.3.3",
    "coreschema==0.0.4",
    "crispy-bootstrap5==2024.2",
    "cryptography==42.0.5",
    "daphne==4.1.2",
    "defusedxml==0.7.1",
    "dj-rest-auth==6.0.0",
    "Django==5.0.4",
    "django-allauth==65.0.2",
    "django-autoslug==1.9.9",
    "django-countries==7.6.1",
    "django-crispy-forms==2.1",
    "django-extensions==3.2.3",
    "django-filter==24.2",
    "django-jazzmin==3.0.0",
    "django-oauth-toolkit==2.4.0",
    "django-redis==5.4.0",
    "django-rest-framework-social-oauth2==1.2.0",
    "django-templated-mail==1.1.1",
    "djangorestframework==3.15.1",
    "docopt==0.6.2",
    "hyperlink==21.0.0",
    "idna==3.7",
    "incremental==24.7.2",
    "itypes==1.2.0",
    "Jinja2==3.1.3",
    "jwcrypto==1.5.6",
    "MarkupSafe==2.1.5",
    "msgpack==1.1.0",
    "oauthlib==3.2.2",
    "packaging==24.0",
    "pillow==10.3.0",
    "pip-upgrader==1.4.15",
    "psycopg2==2.9.10",
    "psycopg2-binary==2.9.9",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pycparser==2.22",
    "PyJWT==2.8.0",
    "pyOpenSSL==24.2.1",
    "python-dotenv==1.0.1",
    "python3-openid==3.2.0",
    "pytz==2024.1",
    "redis==5.1.1",
    "requests==2.31.0",
    "requests-oauthlib==2.0.0",
    "service-identity==24.1.0",
    "six==1.16.0",
    "social-auth-app-django==5.4.0",
    "social-auth-core==4.5.3",
    "sqlparse==0.5.0",
    "terminaltables==3.1.10",
    "tomli==2.0.2",
    "Twisted==24.7.0",
    "txaio==23.1.1",
    "typing_extensions==4.11.0",
    "uritemplate==4.1.1",
    "urllib3==2.2.1",
    "zope.interface==7.1.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"